HTTPClient: Delay AsyncHTTPClient creation.
The AsyncHTTPClient initializer grabs a handle on the active IOLoop
which causes problems if you create a client before you start the IOLoop
and you are running in multi-process mode.  This is easy enough to fix
in this case, so I'll delay grabbing a client until I actually need one.